Select and Query Exercise
In this exercise we cover the basics on selecting and extracting data
using queries. We add a few other materials into it. This should prepare
you for the next exercise: Basic Decision Tree. The purpose of this is
to give you the 20 percent of the tools to get 80 percent of work done.
You will be off and running plotting simple decision trees. It will be
helpful to read up on str(),colnames() and data.frame() command.